fixed 2 more collateral charge causes

+++ src/wallet.h
@@ -206,7 +206,7 @@ public:
     bool CommitTransaction(CWalletTx& wtxNew, CReserveKey& reservekey);
     std::string SendMoney(CScript scriptPubKey, int64 nValue, CWalletTx& wtxNew, bool fAskFee=false, AvailableCoinsType coin_type=ALL_COINS);
     std::string SendMoneyToDestination(const CTxDestination &address, int64 nValue, CWalletTx& wtxNew, bool fAskFee=false, AvailableCoinsType coin_type=ALL_COINS);
-    std::string PrepareDarksendDenominate(int minRounds, int maxAmount);
+    std::string PrepareDarksendDenominate(int minRounds, int64 maxAmount);
     bool CreateCollateralTransaction(CTransaction& txCollateral, std::string strReason);
 
     bool NewKeyPool();
